Rename namespace and references from Mischellaneous to Miscellaneous
